fixing bugs in previous check-in
git-svn-id: file:///Users/sramsey/Desktop/DizzyRepo/ISBJava_Dizzy@2358 5021c4a6-a05a-4b22-8634-7b27b77db61f

+++ ISBJava_Dizzy/trunk/bolouri/java/org/systemsbiology/chem/app/SimulationLauncher.java
@@ -658,7 +658,7 @@ public class SimulationLauncher
             {
                 simulator = (ISimulator) getSimulatorRegistry().getInstance(simulatorAlias);
                 SimulatorParameters simParams = simulator.getDefaultSimulatorParameters();
-                Integer ensembleSize = simParams.getEnsembleSize();
+                Long ensembleSize = simParams.getEnsembleSize();
                 if(null != ensembleSize)
                 {
                     mEnsembleField.setText(ensembleSize.toString());
@@ -672,7 +672,7 @@ public class SimulationLauncher
                     mEnsembleFieldLabel.setEnabled(false);
                 }
                 
-                Integer minNumSteps = simParams.getMinNumSteps();
+                Long minNumSteps = simParams.getMinNumSteps();
                 if(null != minNumSteps)
                 {
                     mNumStepsField.setText(minNumSteps.toString());
